Integrate custom link behavior for MUI components
